+++ crates/runner-nordic/src/main.rs
@@ -14,6 +14,7 @@
 
 #![no_std]
 #![no_main]
+#![feature(never_type)]
 #![feature(try_blocks)]
 
 extern crate alloc;